requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
fastapi==0.115.12
h11==0.14.0
idna==3.10
pydantic==2.11.3
pydantic_core==2.33.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
pypdf>=4.2.0
python-docx==0.8.11
langchain==0.1.16
python-magic>=0.4.27  # File type validation
weaviate-client>=3.26.7,<4.0.0
sentence-transformers>=2.7.0
huggingface_hub[hf_xet]>=0.23.0